refactor FABs to remove duplicated code
