+++ adam-rms.code-workspace
@@ -5,7 +5,7 @@
 		}
 	],
 	"settings": {
-		"editor.tabSize": 4,
+		"editor.tabSize": 2,
 		"typescript.preferences.importModuleSpecifier": "relative",
 		"typescript.preferences.quoteStyle": "double",
 		"eslint.workingDirectories": [
