+++ app/src/components/assets/AssetsFAB.tsx
@@ -1,43 +1,21 @@
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import {
-  useIonRouter,
-  useIonToast,
-  IonFab,
-  IonFabButton,
-  IonFabList,
-} from "@ionic/react";
-import { useContext } from "react";
-import { LocationContext } from "../../contexts/utility/LocationContext";
-import GetAssetFromBarcode from "../../utilities/barcode/GetAssetFromBarcode";
+import { IonFab, IonFabButton, IonFabList } from "@ionic/react";
+import AssetLookupFab from "../fabs/AssetLookupFab";
+import LocationFab from "../fabs/LocationFab";
 
+/**
+ * Floating action buttons for Projects
+ * Create components in app\src\components\fabs and import here
+ */
 const AssetsFab = () => {
-  const router = useIonRouter();
-  const [present] = useIonToast();
-  const { location } = useContext(LocationContext);
-
-  //Define actual button functions here, that then call utility functions
-  /**
-   * Redirect to Asset
-   * Scans a barcode and then goes to that asset
-   */
-  const redirectToAsset = async () => {
-    const asset = await GetAssetFromBarcode(location);
-    if (asset) {
-      router.push("/assets/" + asset.assetTypes_id + "/" + asset.assets_id);
-    } else {
-      present("Asset Not found", 2000);
-    }
-  };
-
   return (
     <IonFab vertical="bottom" horizontal="end" slot="fixed">
       <IonFabButton color="light">
         <FontAwesomeIcon icon="info" />
       </IonFabButton>
       <IonFabList side="top">
-        <IonFabButton onClick={redirectToAsset}>
-          <FontAwesomeIcon icon="search" />
-        </IonFabButton>
+        <AssetLookupFab />
+        <LocationFab />
       </IonFabList>
     </IonFab>
   );
